BE/requirements.txt
fastapi
uvicorn[standard]
typesense
openai
python-dotenv
pydantic 